Add helper to invalidate many packages with one removefilter.
The aim is to cut down on time spent running removefilter when
invalidating dependent packages.

This also uses the `-A` option to restrict the architecture rather than
an `Architecture` clause in the filter. I don't think this will do much
for performance but it should cut down on the line length considerably.