Iteratively generate a list of dependents to remove.
This swaps the recursive dependency removeal functions for an iterative
approach which generates a large set of transitive dependencies.

There's no serious advantage to the iterative approach over the
recursive one. I needed to rewrite the transitive recursion to avoid
removing packages along the way and this approach popped out first.
If recursion is preferred to iterating on an unordered queue we can
change it.