Add repository info parser.
Right now this only parses Packages files in order to find
reverse-dependencies of packages in-process instead of invoking reprepro
once per-package.

The approach is definitely not complete. For one, it only looks in the
`main` component. ROS repositories currently only use the main component
so this is suitable for now.

It's likely that this can be made more general purpose and cover more
use-cases in the future but I'm only developing it as far as I need to
to make it go. Currently it finds all packages that reprepro does in a
tiny fraction of the time.