Queue packages to process as a list rather than a set.
The initial reason for using a set was to totally avoid queuing the same
package twice. But as pointed out in review set ordering isn't
deterministic and will make debugging difficult.

Now we use a list and trust that nothing will get added to the list more
than once per the set difference logic.