Exclude non-ros packages from the package list.
This builds on the change in 285dcde to exclude packages that do not
match `ros-*` from consideration. While building the reverse dependency
chain, each package in the list has its dependencies inspected. The
previous change was sufficient to remove non-ros packages from the
dependencies of ros packages but it did not remove non-ros packages from
the list of packages entirely. This meant that non-ros packages that
directly depend on ros packages were still considered.