Remove dead code.
We've replaced the recursive _invalidate_dependents function with a loop
that walks all transitive dependent packages before invalidating them in
one step with invalidate_packages.

Shelling out to reprepro to gather dependencies was bottlenecking this
process so we've replaced the _get_dependent_packages function with the
RepositoryInfo class's get_rdepends.