Use fileext to narrow scope of renames.
The previous commits assumed that the file extension was the only
occurence of '.ddeb' on any one line. This is a relatively safe
assumtion but that just means we'll have forgotten about it by the time
it finally punishes us.

Instead split of the basename from the ddeb extension and then apply the
.deb extension instead. Afterwards use the created filenames for the
line replacement. Which could have been done previously it just didn't
occur to me.