Add the pyside2 source package to the import formula.
Because the source package for the pyside2 collection is called simply `pyside2` it was not imported as was originally intended along with the rest of these packages. Because the source package for shiboken2 is called `shiboken2` that source package was successfully imported previously.

The version of pyside2 which was imported previously is no longer available in the linked PPA so this import should not be re-run unless the PPA is updated or an examination of the differences between the current ROS and PPA packages is performed.