Spike out a change to allow overriding the source package requirement.
Considering the possibility of using this repository for gurumdds
updates as well but since that's a proprietary package redistributed by
agreement it does not have source packages available and thus we would
need to be able to override our own check.

This is a minimal change which adds a new implicitly boolean field but
I'd like to get feedback on further refactoring to allow overrides to be
package specific (although I have no immediate use case for it).

Examples:

Override all source requirements:

```
override_sources_required: true
```

Override requirements for specific packages

```
override_sources_required:
  - gurumdds-2.6
  - gurumdds-2.7
```