Update aptly key (#147)
* Update aptly repo key.

Fetch via https rather than the keyserver network since the key appears
to've changed since we last updated these scripts and this makes it
robust to those changes as long as the aptly devs keep updating their
https-hosted public key. Also, we've seen that the keyserver network is
crumbling and this will probably be more reliable in general.

* Drop trailing whitespace.

Caught automatically by my editor.